CopyObject & compression changes
Oops, forgot to commit again! 😅 We switch to CopyObject in the backup, because I want to stop pulling from S3 for all this stuff anyway to reduce the cost of running this in bulk. But we still read from S3 for now, that change comes later!

Second, I saw an image compress _larger_ than it came in! I moved the min quality threshold down, because I preferred the 25% size over the minor visible dithering, and added logic to handle the failure case in the future.