Generate images ourselves, instead of S3 GET
This is a bit of a money cheat! Instead of reading from S3, which will end up being 1TB of transfer to actually do the whole thing; or reading from `impress-outfit-images.openneo.net`, which will be 1TB of transfer out of Vercel,

instead we're generating the images locally, based on GraphQL requests from `impress-2020.openneo.net`.

This puts additional CPU pressure on my machine for the eventual bulk operation, which is fine; and moves the bandwidth relationship to be between me and the Neopets and Fastly CDNs, which will be much better for my bill, anyway 😅 (CDNs are generally much less bandwidth-constrained than servers that generate stuff!)