Add delete-all-tags script
Huh, well! The tags were a cool way to keep track of what was going on during the backup process, but I'm paying 17 cents a day for the ongoing tag storage, which seems silly!

Here, we copy-paste the basic iteration shell from `backup-all.js`, but we do a simple S3 `deleteObjectTagging` for each image instead.